requirements.txt
Django==2.0
django-cors-headers==2.1.*
djangorestframework==3.7.*
pytz==2017.3
Pillow==4.3.*
coreapi==2.3.*
djangorestframework-simplejwt==3.2
